+++ src/custom/git.bash
@@ -2,7 +2,11 @@
 
 git_add_modified()
 {
-    ga $(gs | grep -Pe "modified:.*" | grep -Poe "(?<=modified:   )([^ ]+)" | tr "\n" " ")
+    local modified_files
+
+    modified_files="$(gs | grep -Pe "modified:.*" | grep -Poe "(?<=modified:   )([^ ]+)" | tr "\n" " ")"
+    # shellcheck disable=SC2086
+    ga ${modified_files}
 }
 
 gax()
@@ -10,6 +14,7 @@ gax()
     if [ $# -eq 0 ]; then
         git add .
     else
+        # shellcheck disable=SC2046
         git add -- . $(printf ":!%s " "$@")
     fi
 }