+++ src/plugins/local-bashrc.bash
@@ -3,10 +3,11 @@
 if [ -z "${BASHRC_INC}" ]; then
     BASHRC_INC=$HOME/.bashrc.d
 fi
-if [[ -d ${BASHRC_INC} ]]; then    
-    for i in ${BASHRC_INC}/*.*sh; do
-        if [ -r $i ]; then
-            . $i
+if [[ -d ${BASHRC_INC} ]]; then
+    for i in "${BASHRC_INC}"/*.*sh; do
+        if [ -r "$i" ]; then
+            # shellcheck source=/dev/null
+            . "$i"
         fi
     done
     unset i