+++ src/plugins/local-completion.bash
@@ -4,9 +4,10 @@ if [ -z "${BASH_COMPLETION_INC}" ]; then
     BASH_COMPLETION_INC=$HOME/.bash_completion.d
 fi
 if [[ -d ${BASH_COMPLETION_INC} ]]; then
-    for i in ${BASH_COMPLETION_INC}/*.*sh; do
-        if [ -r $i ]; then
-            . $i
+    for i in "${BASH_COMPLETION_INC}"/*.*sh; do
+        if [ -r "$i" ]; then
+            # shellcheck source=/dev/null
+            . "$i"
         fi
     done
     unset i