+++ updates/v0.1.1.sh
@@ -1,3 +1,4 @@
+# shellcheck shell=bash
 update_v011()
 {
     [ -e "${BASH_IT}/lib/custom.bash" ] && unlink "${BASH_IT}/lib/custom.bash"